requirements.txt
gymnasium==1.2.3
gymnasium[mujoco]==1.2.3
pandas==2.3.3
matplotlib==3.10.8
numpy==2.4.4
omegaconf==2.3.0
PyYAML==6.0.3
torch==2.11.0
tensorboard==2.20.0
mpi4py==4.1.1
